refactor formbuilder base
